removed dupes + redefined incorrect models
