+++ weather_bites/requirements.txt
@@ -1,96 +1,22 @@
-annotated-types==0.7.0
-asgiref==3.7.2
 blinker==1.8.2
-blis==0.7.11
-cachetools==5.4.0
-catalogue==2.0.10
+build==1.2.2.post1
 certifi==2023.11.17
 charset-normalizer==3.3.2
 click==8.1.7
-cloudpathlib==0.18.1
-confection==0.1.5
-contourpy==1.2.0
-cycler==0.12.1
-cymem==2.0.8
-Django==4.2.4
-docstring_parser==0.16
 Flask==3.0.3
-fonttools==4.46.0
-google-api-core==2.19.1
-google-auth==2.32.0
-google-cloud-aiplatform==1.59.0
-google-cloud-bigquery==3.25.0
-google-cloud-core==2.4.1
-google-cloud-resource-manager==1.12.4
-google-cloud-storage==2.17.0
-google-cloud-vision==3.7.3
-google-crc32c==1.5.0
-google-resumable-media==2.7.1
-googleapis-common-protos==1.63.2
-grpc-google-iam-v1==0.13.1
-grpcio==1.64.1
-grpcio-status==1.62.2
+Flask-SQLAlchemy==3.1.1
 idna==3.7
-iniconfig==2.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.4
-joblib==1.3.2
-kiwisolver==1.4.5
-langcodes==3.4.0
-language_data==1.2.0
-marisa-trie==1.2.0
-markdown-it-py==3.0.0
 MarkupSafe==2.1.5
-matplotlib==3.8.2
-mdurl==0.1.2
-mocker==1.1.1
-murmurhash==1.0.10
-mysql-connector-python==9.0.0
-networkx==3.2.1
-numpy==1.26.2
-packaging==23.2
-patch==1.16
-pdf2image==1.17.0
-Pillow==10.1.0
-pluggy==1.5.0
-preshed==3.0.9
-proto-plus==1.24.0
-protobuf==4.25.3
-pyasn1==0.6.0
-pyasn1_modules==0.4.0
-pydantic==2.8.2
-pydantic_core==2.20.1
-Pygments==2.18.0
-PyMuPDF==1.24.5
-PyMuPDFb==1.24.3
-pyparsing==3.1.1
-pytesseract==0.3.10
-pytest==8.3.3
-pytest-mock==3.14.0
-python-dateutil==2.8.2
+packaging==24.2
+pip-tools==7.4.1
+pyproject_hooks==1.2.0
 python-dotenv==1.0.1
 requests==2.32.3
-rich==13.7.1
-rsa==4.9
-scikit-learn==1.3.2
-scipy==1.11.4
-shapely==2.0.5
-shellingham==1.5.4
-six==1.16.0
-sklearn==0.0.post11
-smart-open==7.0.4
-spacy==3.7.5
-spacy-legacy==3.0.12
-spacy-loggers==1.0.5
-sqlparse==0.4.4
-srsly==2.4.8
-thinc==8.2.5
-threadpoolctl==3.2.0
-tqdm==4.66.4
-typer==0.12.3
+setuptools==75.6.0
+SQLAlchemy==2.0.36
 typing_extensions==4.12.2
 urllib3==2.2.2
-wasabi==1.1.3
-weasel==0.4.1
 Werkzeug==3.0.4
-wrapt==1.16.0
+wheel==0.45.1
