+++ weather_bites/weather_bites/models/db.py
@@ -1,10 +1,10 @@
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
+from werkzeug.security import generate_password_hash, check_password_hash
 
 # Initialize the database
 db = SQLAlchemy()
 
-# User model
 class User(db.Model):
     __tablename__ = 'users'
 
@@ -15,6 +15,14 @@ class User(db.Model):
 
     # Relationships
     favorites = db.relationship('Favorite', backref='user', lazy=True)
+
+    def set_password(self, password):
+        """Hashes the password and sets the password_hash."""
+        self.password_hash = generate_password_hash(password)
+
+    def check_password(self, password):
+        """Checks the hashed password."""
+        return check_password_hash(self.password_hash, password)
 
     def __repr__(self):
         return f'<User {self.username}>'
@@ -30,3 +38,18 @@ class Favorite(db.Model):
 
     def __repr__(self):
         return f'<Favorite {self.city} for User ID {self.user_id}>'
+
+# Review model
+class Review(db.Model):
+    __tablename__ = 'reviews'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String, nullable=False, unique=True)
+    location = db.Column(db.String, nullable=False)
+    rating = db.Column(db.Integer, nullable=False)
+    favorite = db.Column(db.Boolean, default=False)
+    review = db.Column(db.Text, nullable=True)
+    deleted = db.Column(db.Boolean, default=False)
+
+    def mark_deleted(self):
+        self.deleted = True
